fix: calculate result value when change from-unit or to-unit
